move sge def and functions into separate header for further extension
