+++ backend/common/dbbe_api.h
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2018,2019 IBM Corporation
+ * Copyright © 2018-2020 IBM Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,11 +27,12 @@
  * @{
  */
 
+#include "common/sge.h"
+
 #include <inttypes.h>
 #include <stddef.h> // size_t
 #include <stdlib.h> // calloc
 #include <libdatabroker.h>
-#include <sys/uio.h>
 
 /**
  * @typedef dbBE_Handle_t
@@ -52,21 +53,6 @@ typedef void* dbBE_Handle_t;
  * subsequent API calls when accessing a namespace.
  */
 typedef void* dbBE_NS_Handle_t;
-
-/*
- * max number of SGEs in assembled redis commands (IOV_MAX replacement)
- */
-#define DBBE_SGE_MAX ( 256 )
-
-/**
- *  @struct dbBE_sge_t dbbe_api.h "backend/common/dbbe_api.h"
- *
- *  @brief Scatter-Gather-Element structure
- *
- *  The scatter-gather-element defines one entry in a scatter-gather list
- *  and specifies one contiguous memory region in main memory.
- */
-typedef struct iovec dbBE_sge_t;
 
 /**
  * @brief Back-end operation code
@@ -537,28 +523,6 @@ extern const dbBE_api_t g_dbBE;
  * A set of helper functions to handle requests, completions, or SGEs
  */
 
-
-/**
- * @brief calculate the total size of an SGE list
- *
- * Browses the SGEs of an input list and returns the total size in bytes.
- *
- * @param [in] sge       pointer to the first entry of an SGE list (array)
- * @param [in] sge_count number of elements in the SGE list
- *
- * @return size of the SGE in bytes
- */
-static inline
-size_t dbBE_SGE_get_len( const dbBE_sge_t *sge, const int sge_count )
-{
-  if( sge == NULL )
-    return 0;
-  int i;
-  size_t len = 0;
-  for( i = sge_count-1; i>=0; --i )
-    len += sge[ i ].iov_len;
-  return len;
-}
 
 /**
  * @brief allocate a new request struct including SGE-space
