+++ backend/common/sge.h
@@ -1,0 +1,65 @@
+/*
+ * Copyright © 2020 IBM Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+#ifndef BACKEND_COMMON_SGE_H_
+#define BACKEND_COMMON_SGE_H_
+
+
+#include <sys/uio.h> // struct iovec
+#include <stddef.h> // NULL
+
+
+/*
+ * max number of SGEs in assembled redis commands (IOV_MAX replacement)
+ */
+#define DBBE_SGE_MAX ( 256 )
+
+/**
+ *  @struct dbBE_sge_t dbbe_api.h "backend/common/dbbe_api.h"
+ *
+ *  @brief Scatter-Gather-Element structure
+ *
+ *  The scatter-gather-element defines one entry in a scatter-gather list
+ *  and specifies one contiguous memory region in main memory.
+ */
+typedef struct iovec dbBE_sge_t;
+
+
+
+/**
+ * @brief calculate the total size of an SGE list
+ *
+ * Browses the SGEs of an input list and returns the total size in bytes.
+ *
+ * @param [in] sge       pointer to the first entry of an SGE list (array)
+ * @param [in] sge_count number of elements in the SGE list
+ *
+ * @return size of the SGE in bytes
+ */
+static inline
+size_t dbBE_SGE_get_len( const dbBE_sge_t *sge, const int sge_count )
+{
+  if( sge == NULL )
+    return 0;
+  int i;
+  size_t len = 0;
+  for( i = sge_count-1; i>=0; --i )
+    len += sge[ i ].iov_len;
+  return len;
+}
+
+#endif /* BACKEND_COMMON_SGE_H_ */
